--- client/task/task.go
@@ -1,10 +1,0 @@
-package task
-
-import (
-	"context"
-)
-
-// Task to be performed.
-type Task interface {
-	Perform(context.Context) ([]string, error)
-}
