+++ client/v1/transport/grpc/client.go
@@ -1,8 +1,15 @@
 package grpc
 
 import (
+	"context"
+
+	sgrpc "github.com/alexfalkowski/go-service/transport/grpc"
+	"github.com/alexfalkowski/go-service/transport/grpc/metrics/prometheus"
+	"github.com/alexfalkowski/go-service/transport/grpc/trace/opentracing"
 	v1 "github.com/alexfalkowski/migrieren/api/migrieren/v1"
+	v1c "github.com/alexfalkowski/migrieren/client/v1/config"
 	"go.uber.org/fx"
+	"go.uber.org/zap"
 	"google.golang.org/grpc"
 )
 
@@ -10,10 +17,33 @@ import (
 type ServiceClientParams struct {
 	fx.In
 
-	Conn *grpc.ClientConn
+	Lifecycle fx.Lifecycle
+	Config    *sgrpc.Config
+	Logger    *zap.Logger
+	Tracer    opentracing.Tracer
+	Client    *v1c.Config
+	Metrics   *prometheus.ClientMetrics
 }
 
 // NewServiceClient for gRPC.
-func NewServiceClient(params ServiceClientParams) v1.ServiceClient {
-	return v1.NewServiceClient(params.Conn)
+func NewServiceClient(params ServiceClientParams) (v1.ServiceClient, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), params.Client.Timeout)
+	defer cancel()
+
+	conn, err := sgrpc.NewClient(
+		sgrpc.ClientParams{Context: ctx, Host: params.Client.Host, Config: params.Config},
+		sgrpc.WithClientLogger(params.Logger), sgrpc.WithClientTracer(params.Tracer), sgrpc.WithClientDialOption(grpc.WithBlock()),
+		sgrpc.WithClientMetrics(params.Metrics),
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	params.Lifecycle.Append(fx.Hook{
+		OnStop: func(ctx context.Context) error {
+			return conn.Close()
+		},
+	})
+
+	return v1.NewServiceClient(conn), nil
 }
