+++ cmd/client.go
@@ -5,7 +5,7 @@ import (
 	"github.com/alexfalkowski/go-service/logger"
 	"github.com/alexfalkowski/go-service/marshaller"
 	"github.com/alexfalkowski/go-service/transport"
-	v1 "github.com/alexfalkowski/migrieren/client/v1"
+	"github.com/alexfalkowski/migrieren/client"
 	"github.com/alexfalkowski/migrieren/config"
 	"go.uber.org/fx"
 )
@@ -15,5 +15,5 @@ var ClientOptions = []fx.Option{
 	fx.NopLogger,
 	marshaller.Module, cmd.Module,
 	fx.Provide(NewVersion), config.Module, logger.ZapModule,
-	transport.GRPCModule, v1.Module,
+	transport.GRPCModule, client.Module,
 }
